+++ ReadingWriting.h
@@ -8,7 +8,7 @@ using namespace std;
 int file_read( double *nx,double *ny, double *nz, double *px, double *py,double*pz);
 int file_read_ASCII( double *nx,double *ny, double *nz, double *px, double *py,double*pz);
 
-void writeVTKfiles(double n,double *nx,double *ny,double *nz,double *px,double *py,double *pz) ;
-void writeBENDfiles(double n, double* nx,double* ny,double* nz) ;
+void writeVTKfiles(int n,double *nx,double *ny,double *nz,double *px,double *py,double *pz) ;
+void writeBENDfiles(int n, double* nx,double* ny,double* nz) ;
 
 #endif //READINGWRITING_H
